Rename/restructure visualizer files
