+++ packages/app/src/vis-packs/VisBoundary.tsx
@@ -2,9 +2,9 @@ import type { ReactNode } from 'react';
 import { Suspense, useContext } from 'react';
 import { ErrorBoundary } from 'react-error-boundary';
 
+import ErrorFallback from '../ErrorFallback';
 import { ProviderContext } from '../providers/context';
-import ErrorFallback from '../visualizer/ErrorFallback';
-import ValueLoader from '../visualizer/ValueLoader';
+import ValueLoader from './ValueLoader';
 
 interface Props {
   resetKey?: unknown;
