+++ packages/app/src/vis-packs/core/complex/ComplexLineVisContainer.tsx
@@ -5,8 +5,8 @@ import {
 } from '@h5web/shared';
 
 import DimensionMapper from '../../../dimension-mapper/DimensionMapper';
+import { useDimMappingState } from '../../../dimension-mapper/hooks';
 import VisBoundary from '../../VisBoundary';
-import { useDimMappingState } from '../../hooks';
 import type { VisContainerProps } from '../../models';
 import ValueFetcher from '../ValueFetcher';
 import MappedComplexLineVis from './MappedComplexLineVis';
