+++ packages/app/src/vis-packs/core/complex/ComplexVisContainer.tsx
@@ -6,8 +6,8 @@ import {
 } from '@h5web/shared';
 
 import DimensionMapper from '../../../dimension-mapper/DimensionMapper';
+import { useDimMappingState } from '../../../dimension-mapper/hooks';
 import VisBoundary from '../../VisBoundary';
-import { useDimMappingState } from '../../hooks';
 import type { VisContainerProps } from '../../models';
 import ValueFetcher from '../ValueFetcher';
 import MappedComplexVis from './MappedComplexVis';
