+++ packages/app/src/vis-packs/nexus/containers/NxComplexImageContainer.tsx
@@ -1,9 +1,9 @@
 import { assertGroupWithChildren, assertMinDims } from '@h5web/shared';
 
 import DimensionMapper from '../../../dimension-mapper/DimensionMapper';
+import { useDimMappingState } from '../../../dimension-mapper/hooks';
 import VisBoundary from '../../VisBoundary';
 import MappedComplexVis from '../../core/complex/MappedComplexVis';
-import { useDimMappingState } from '../../hooks';
 import type { VisContainerProps } from '../../models';
 import NxValuesFetcher from '../NxValuesFetcher';
 import { getNxData, getDatasetLabel, assertComplexSignal } from '../utils';
