+++ packages/app/src/visualizer/Visualizer.module.css
@@ -1,15 +1,9 @@
-.visualizer {
+.visManager {
   flex: 1 1 0%;
   min-height: 0;
   display: flex;
   flex-direction: column;
   scrollbar-width: thin;
-}
-
-.displayArea {
-  flex: 1 1 0%;
-  min-height: 0;
-  display: flex;
 }
 
 .visBar {
@@ -21,18 +15,12 @@
   background-color: var(--secondary-light-bg);
 }
 
+.displayArea {
+  flex: 1 1 0%;
+  min-height: 0;
+  display: flex;
+}
+
 .fallback {
   composes: fallback from global;
-}
-
-.error {
-  composes: error from global;
-}
-
-.error > span {
-  margin-left: 0.5rem;
-}
-
-.retryBtn {
-  composes: btnClean btnLink from global;
 }
